refactor: introduce Syscall interface and consolidate wrappers
Add a single Syscall interface that domain code goes through, with
LinuxSyscall as the production implementation invoking platform.posix.* /
platform.linux.* calls. This is the testability seam: future PRs will
add a fake Syscall implementation so domain logic can be unit-tested
without invoking real kernel calls.

Existing per-syscall files (Mount.kt, Rlimit.kt, Prctl.kt, CloseRange.kt,
Groups.kt, Signal.kt#killProcess) collapse into LinuxSyscall methods.
The unused setDumpable / getDumpable helpers in Prctl.kt are dropped (no
callers since the bootstrap.c refactor). Existing call sites in rootfs/,
process/, command/Kill.kt, command/Delete.kt now go through a temporary
top-level `defaultSyscall` binding so callers don't yet need a parameter.
That binding is intentionally short-lived and will be removed in the next
step when each domain function takes a Syscall parameter explicitly.

No functional change — every wrapper preserves its existing behavior
(close_range fallback, /proc/self/setgroups check, ESRCH-on-kill swallow,
rlimit-type-to-resource mapping, etc.).
